refactor: Service 클래스 재사용
