+++ src/test/java/nextstep/web/RoomEscapeServiceTest.java
@@ -7,6 +7,7 @@ import java.time.LocalTime;
 import nextstep.exception.ReservationDuplicateException;
 import nextstep.exception.ReservationNotFoundException;
 import nextstep.repository.JdbcTemplateReservationRepository;
+import nextstep.service.RoomEscapeService;
 import nextstep.web.dto.ReservationRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
